Add itemsSelling and allItemsSelling()
Major change 1:
Added a variable called itemsSelling which works similarly to itemsSold
and itemsBought, in that it stores all users as properties,
and each property contains a list of listingIDs which point to items
being currently sold. However that list is an object rather than
an array, and each property is stored as "listingID:true"
or "listingID:false", in contrast to the other lists.

 Don't worry about it though, as the newly added function
 "allItemsSelling(sellerID)" will still return an array of
 only those items which are currently being sold by that seller.

 Major change 2:
 changed GetItemDescription's functionality
 so that it will return the full object corresponding to the listingID,
 rather than just an object with price, title and blurb.
 This is as per your request, I hope that it works as expected.